Move render-component to a separate file
