+++ src/components/navbar/navbar.tsx
@@ -8,7 +8,7 @@ import type * as normalizePages from "nextra/normalize-pages"
 import { Fragment, useState, type ReactElement, type ReactNode } from "react"
 import { useMenu, useThemeConfig } from "nextra-theme-docs"
 import { Anchor } from "@/app/conf/_design-system/anchor"
-import { renderComponent } from "@/components/utils"
+import { renderComponent } from "@/components/utils/render-component"
 
 import MenuIcon from "@/app/conf/_design-system/pixelarticons/menu.svg?svgr"
 import CloseIcon from "@/app/conf/_design-system/pixelarticons/close.svg?svgr"
